--- router/pool/rateLimiter_test.go
@@ -1,1 +1,0 @@
-package pool
